services/technical_indicators/pyproject.toml
[project]
name = "technical-indicators"
version = "0.1.0"
description = "Technical indicators calculation service for real-time ML trading system"
readme = "README.md"
authors = [
    { name = "WilliamJudge94", email = "37912237+WilliamJudge94@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "numpy",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "psycopg2-binary",
    "quixstreams>=3.22.0",
    "ta-lib",
]

[project.scripts]
technical_indicators = "technical_indicators:main"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
